sedmi commit
Glavna izmena brisanje sa brisanjem slika u folderu i jos neke ispravke

+++ src/main/java/com/it355/MladenStolicProjekat/controller/AccomodationController.java
@@ -1,12 +1,18 @@
 package com.it355.MladenStolicProjekat.controller;
 
 import com.it355.MladenStolicProjekat.entity.Accommodation;
+import com.it355.MladenStolicProjekat.entity.Accommodationphoto;
+import com.it355.MladenStolicProjekat.service.AccommodationphotoService;
 import com.it355.MladenStolicProjekat.service.AccomodationService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -18,6 +24,7 @@ import java.util.Optional;
 public class AccomodationController {
 
     final AccomodationService accomodationService;
+    final AccommodationphotoService accommodationphotoService;
 
     @GetMapping("")
     public ResponseEntity<List<Accommodation>> getAllAccommodations() {
@@ -70,12 +77,29 @@ public class AccomodationController {
         }
     }
 
-
+    private static final String UPLOAD_DIR = "./src/main/resources/static/images/";
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteAccommodation(@PathVariable int id) {
-        accomodationService.deleteAccommodationById(id);
-        return ResponseEntity.ok().build();
+        Optional<?> opt = accomodationService.findById(id);
+
+        if (opt.isPresent()) {
+            Accommodation smestaj = (Accommodation) opt.get();
+            List<Accommodationphoto> list = accommodationphotoService.findAllByAccommodationId(smestaj.getId());
+            for (Accommodationphoto slika : list) {
+                Path target = Paths.get(UPLOAD_DIR + slika.getImageUrl());
+                try {
+                    Files.deleteIfExists(target);
+                    accommodationphotoService.delete(slika.getId());
+                } catch (IOException e) {
+                    return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Map.of("message", e.getMessage()));
+                }
+            }
+            accomodationService.deleteAccommodationById(id);
+            return ResponseEntity.ok(smestaj);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Map.of("message", "Not found"));
     }
+
     @PutMapping("/{id}/image")
     public ResponseEntity<Accommodation> updateAccommodationImage(@PathVariable int id, @RequestParam("imageUrl") String imageUrl, @RequestParam("type") int type) {
         Optional<Accommodation> existingAccommodation = accomodationService.findById(id);
